Fix puppetmaster configuration.
The [master] block was not correctly placed.  The result was that puppet
client started on the same server as the master tried to connect to the
database and failed to load new facts, giving this kind of error:

err: Could not retrieve catalog from remote server: Error 400 on SERVER:
$concat_basedir not defined. Try running again with pluginsync enabled
at

This is now fixed.